Added more tests for Tenor. Some .classpath etc where referencing non-existing directories.

+++ datecalc-joda/src/main/java/net/objectlab/kit/datecalc/joda/LocalDatePeriodCountCalculator.java
@@ -25,12 +25,13 @@ import org.joda.time.Period;
 import org.joda.time.PeriodType;
 
 /**
- * Joda <code>LocalDatePeriod</code> based implementation of the {@link PeriodCountCalculator}
- *
+ * Joda <code>LocalDatePeriod</code> based implementation of the
+ * {@link PeriodCountCalculator}
+ * 
  * @author Benoit Xhenseval
  * @author $LastChangedBy: marchy $
  * @version $Revision: 138 $ $Date: 2006-09-10 14:29:15 +0200 (Sun, 10 Sep 2006) $
- *
+ * 
  */
 public class LocalDatePeriodCountCalculator implements PeriodCountCalculator<LocalDate> {
 
@@ -98,8 +99,8 @@ public class LocalDatePeriodCountCalculator implements PeriodCountCalculator<Loc
 
                 final int diff1 = new Period(start, endOfStartYear, PeriodType.days()).getDays();
                 final int diff2 = new Period(startOfEndYear, end, PeriodType.days()).getDays();
-                diff = ((double) (diff1 + 1.0)) / (double) start.dayOfYear().getMaximumValue()
-                        + ((double) (endYear - startYear - 1.0)) + ((double) (diff2)) / (double) end.dayOfYear().getMaximumValue();
+                diff = ((diff1 + 1.0)) / start.dayOfYear().getMaximumValue()
+                        + ((endYear - startYear - 1.0)) + ((double) (diff2)) / (double) end.dayOfYear().getMaximumValue();
             }
 
         } else if (basis == PeriodCountBasis.CONV_30_360 || basis == PeriodCountBasis.CONV_360E_ISDA
